# go-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  entities.AddToCartRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  entities.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/entities.PaginationResponse'
      success:
        type: boolean
    type: object
  entities.Cart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/entities.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entities.CartItem:
    properties:
      cart_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/entities.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  entities.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.CreateCategoryRequest:
    properties:
      description:
        type: string
      image:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  entities.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  entities.Product:
    properties:
      category:
        $ref: '#/definitions/entities.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      images:
        items:
          $ref: '#/definitions/entities.ProductImage'
        type: array
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  entities.ProductImage:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      product_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
  entities.UpdateCategoryRequest:
    properties:
      description:
        type: string
      image:
        type: string
      name:
        type: string
    type: object
...
host: localhost:4001
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: API สำหรับระบบ E-commerce ที่พัฒนาด้วย Go Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go E-commerce API
  version: 1.0.0
paths:
...
  /cart:
    delete:
      consumes:
      - application/json
      description: ลบสินค้าทั้งหมดออกจากตะกร้าสินค้า
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ล้างตะกร้าสินค้า
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: ดูตะกร้าสินค้าของผู้ใช้ปัจจุบัน
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Cart'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ดูตะกร้าสินค้า
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: เพิ่มสินค้าลงในตะกร้าสินค้า
      parameters:
      - description: ข้อมูลการเพิ่มสินค้าลงตะกร้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: เพิ่มสินค้าลงตะกร้า
      tags:
      - Cart
  /cart/{itemId}:
    delete:
      consumes:
      - application/json
      description: ลบสินค้าออกจากตะกร้าสินค้า
      parameters:
      - description: Cart Item ID
        in: path
        name: itemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ลบสินค้าจากตะกร้า
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: อัพเดทจำนวนสินค้าในตะกร้า
      parameters:
      - description: Cart Item ID
        in: path
        name: itemId
        required: true
        type: string
      - description: ข้อมูลการอัพเดทสินค้าในตะกร้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: อัพเดทสินค้าในตะกร้า
      tags:
      - Cart
  /categories:
    get:
      consumes:
      - application/json
      description: ดูหมวดหมู่ทั้งหมดพร้อม pagination
      parameters:
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Category'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      summary: ดูหมวดหมู่ทั้งหมด
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: สร้างหมวดหมู่ใหม่ (เฉพาะ Admin)
      parameters:
      - description: ข้อมูลการสร้างหมวดหมู่
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: สร้างหมวดหมู่
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: ลบหมวดหมู่ตาม ID (เฉพาะ Admin)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ลบหมวดหมู่
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: ดูรายละเอียดหมวดหมู่ตาม ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      summary: ดูหมวดหมู่ตาม ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: แก้ไขข้อมูลหมวดหมู่ (เฉพาะ Admin)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการแก้ไขหมวดหมู่
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขหมวดหมู่
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token สำหรับการยืนยันตัวตน ให้ใส่ token ในรูปแบบ: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"